requirements.txt
tensorflow==2.12.0
fastapi==0.95.2
uvicorn==0.22.0
pillow==9.5.0
numpy==1.24.3
flask==2.2.3
